Make CompleteHistory the default and put type in separate file
